+++ tests/compiler/test_compile_export.py
@@ -25,29 +25,29 @@ import pytest
 import inmanta.ast.export as ast_export
 import inmanta.compiler as compiler
 from inmanta.ast import CompilerException, DoubleSetException, ExplicitPluginException, NotFoundException
-from inmanta.compiler.data import ExportCompileData
 from inmanta.config import Config
+from inmanta.data.model import CompileData
 from inmanta.parser import ParserException
 
 
-def export_to_file(path: Path, expected_error_type: Optional[Type[CompilerException]] = None) -> ExportCompileData:
+def export_to_file(path: Path, expected_error_type: Optional[Type[CompilerException]] = None) -> CompileData:
     """
-        Compiles, exporting to a file, and returns the file contents, loaded as ExportCompileData.
+        Compiles, exporting to a file, and returns the file contents, loaded as CompileData.
     """
-    Config.set("compiler", "json", "true")
-    Config.set("compiler", "json_file", str(path))
+    Config.set("compiler", "export_compile_data", "true")
+    Config.set("compiler", "export_compile_data_file", str(path))
     if expected_error_type is None:
         compiler.do_compile()
     else:
         with pytest.raises(CompilerException):
             compiler.do_compile()
     with path.open() as f:
-        return ExportCompileData(**json.loads(f.read()))
+        return CompileData(**json.loads(f.read()))
 
 
 @pytest.fixture
-def tempfile_export(tmp_path: Path) -> Callable[[Optional[Type[CompilerException]]], ExportCompileData]:
-    def do_export(expected_error_type: Optional[Type[CompilerException]] = None) -> ExportCompileData:
+def tempfile_export(tmp_path: Path) -> Callable[[Optional[Type[CompilerException]]], CompileData]:
+    def do_export(expected_error_type: Optional[Type[CompilerException]] = None) -> CompileData:
         return export_to_file(tmp_path / "myfile", expected_error_type)
 
     return do_export
@@ -80,7 +80,7 @@ x = 0
 x = 1
         """,
     )
-    compile_data: ExportCompileData = tempfile_export(DoubleSetException)
+    compile_data: CompileData = tempfile_export(DoubleSetException)
     assert len(compile_data.errors) == 1
     error: ast_export.Error = compile_data.errors[0]
     assert error.category == ast_export.ErrorCategory.runtime
@@ -107,6 +107,6 @@ def test_export_compile_data_to_file_categories(
     snippetcompiler, snippet: str, exception: Type[CompilerException], category: ast_export.ErrorCategory, tempfile_export
 ) -> None:
     snippetcompiler.setup_for_snippet(snippet)
-    compile_data: ExportCompileData = tempfile_export(exception)
+    compile_data: CompileData = tempfile_export(exception)
     assert len(compile_data.errors) == 1
     assert compile_data.errors[0].category == category